Uplift trait RecordFields to tracing-core
For the time being, it's also been unsealed. There doesn't seem to be a
specific reason to seal it, other than it just existing to generalize
over the provided types, and not being an input generalization.

It can be resealed without issue; however, it being in tracing-core is
necessary for what's coming up.
