+++ tracing-core/src/span.rs
@@ -211,6 +211,12 @@ impl<'a> Attributes<'a> {
     }
 }
 
+impl<'a> field::RecordFields<'a> for Attributes<'a> {
+    fn record(&self, visitor: &mut dyn field::Visit<'a>) {
+        Attributes::record(self, visitor)
+    }
+}
+
 // ===== impl Record =====
 
 impl<'a> Record<'a> {
@@ -234,6 +240,12 @@ impl<'a> Record<'a> {
     /// Returns true if this `Record` contains _no_ values.
     pub fn is_empty(&self) -> bool {
         self.values.is_empty()
+    }
+}
+
+impl<'a> field::RecordFields<'a> for Record<'a> {
+    fn record(&self, visitor: &mut dyn field::Visit<'a>) {
+        Record::record(self, visitor)
     }
 }
